Use jekyll
Due to Jekyll's directory structure events have to be in the _posts
folder. They should be named according to their date, e.g.:
"2013-09-12-rejectjs.html"

Events are ordered ascending by their date, which means the event that
is closest to the current date (= the date when jekyll builds the site,
not the actual current date) will be on top of the list.